+++ Spark/src/Core/SparkleRel/Application.cpp
@@ -2,6 +2,8 @@
 
 #include "Application.h"
 
+#include "Event/ApplicationEvent.h"
+
 namespace Spark {
 
 	Application::Application()
@@ -15,8 +17,12 @@ namespace Spark {
 
 	void Application::tickUpdate()
 	{
+		WindowResizedEvent e(1280, 720);
+		SP_CORE_LOG_TRACE(e.ToString())
+
 		while (m_Running)
 		{
+
 		}
 	}
 }