+++ Spark/src/Core/SparkleRel/MainEntry.h
@@ -14,6 +14,7 @@ int main()
 
 	/* --- Test ground --- */
 
+
 	SP_CORE_LOG_INFO("Hello!")
 	SP_LOG_INFO("Hello!")